docs(table): improve source rendering
docs(table): improve source rendering
