+++ packages/core/src/components/Table/stories/Table.stories.tsx
@@ -1,6 +1,4 @@
 import { useMemo } from "react";
-import styled from "@emotion/styled";
-import { useTheme, Breakpoints as MuiBreakpoints } from "@mui/material/styles";
 import { Ban } from "@hitachivantara/uikit-react-icons";
 import { StoryObj } from "@storybook/react";
 import {
@@ -17,112 +15,21 @@ import {
   HvCheckBox,
   HvTypography,
   HvDropDownMenu,
-  HvButton,
-  HvOverflowTooltip,
-  theme,
 } from "@hitachivantara/uikit-react-core";
 
 import {
   AssetEvent,
   getColumns,
-  getGroupedRowsColumns,
   makeData,
   useToggleIndex,
 } from "./storiesUtils";
 
-// #region Responsive table styled components
-
-const StyledResponsiveTableContainer = styled(HvTableContainer)({
-  inWidth: 200,
-  maxHeight: 300,
-});
-
-const StyledResponsiveTable = styled(HvTable)({
-  display: "flex",
-  flexFlow: "column wrap",
-});
-
-const StyledResponsiveHead = styled(HvTableHead)(
-  ({ $breakpoints }: { $breakpoints: MuiBreakpoints }) => ({
-    display: "flex",
-    flexFlow: "column wrap",
-    height: "auto",
-
-    [$breakpoints.only("md")]: {
-      "&:first-of-type": {
-        position: "sticky",
-        top: -1,
-        zIndex: 3,
-      },
-    },
-
-    [$breakpoints.down("sm")]: {
-      "&:first-of-type": {
-        display: "none",
-      },
-    },
-  })
-);
-
-const StyledResponsiveBody = styled(HvTableBody)(
-  ({ $breakpoints }: { $breakpoints: MuiBreakpoints }) => ({
-    display: "flex",
-    flexFlow: "column wrap",
-
-    [$breakpoints.only("md")]: {
-      "&:first-of-type": {
-        position: "sticky",
-        top: -1,
-        zIndex: 3,
-      },
-    },
-
-    [$breakpoints.down("sm")]: {
-      "&:first-of-type": {
-        display: "none",
-      },
-    },
-  })
-);
-
-const StyledResponsiveTableRow = styled(HvTableRow)(
-  ({ $breakpoints }: { $breakpoints: MuiBreakpoints }) => ({
-    display: "flex",
-    flexFlow: "row wrap",
-
-    "&>*": {
-      width: "calc(100% / 7)",
-
-      display: "flex",
-      alignItems: "center",
-
-      [$breakpoints.down("sm")]: {
-        width: "100%",
-
-        "&:first-of-type": {
-          width: "100%",
-          justifyContent: "center",
-          backgroundColor: theme.colors.atmo1,
-        },
-      },
-    },
-
-    [$breakpoints.down("sm")]: {
-      "& > div:not(:first-of-type)::before": {
-        content: "attr(data-label) ",
-        fontWeight: "bold",
-        width: 150,
-      },
-    },
-  })
-);
-
-const StyledResponsiveTableHeader = styled(HvTableHeader)({
-  display: "flex",
-  alignItems: "start",
-});
-
-// #endregion Responsive table styled components
+import { GroupedRows as GroupedRowsStory } from "./TableSamples/GroupedRows";
+import GroupedRowsRaw from "./TableSamples/GroupedRows?raw";
+import { ResponsiveTable as ResponsiveTableStory } from "./TableSamples/ResponsiveTable";
+import ResponsiveTableRaw from "./TableSamples/ResponsiveTable?raw";
+import { ListRow as ListRowStory } from "./TableSamples/ListRow";
+import ListRowRaw from "./TableSamples/ListRow?raw";
 
 export default {
   title: "Visualizations/Table",
@@ -302,111 +209,20 @@ export const SimpleTable: StoryObj<HvTableProps> = {
 export const GroupedRows: StoryObj<HvTableProps> = {
   parameters: {
     docs: {
+      source: { code: GroupedRowsRaw },
       description: {
         story: "A table example with grouped rows.",
       },
     },
   },
-  render: () => {
-    const columns = getGroupedRowsColumns();
-    const data = makeData(8);
-
-    const style = {
-      borderRight: `solid 1px ${theme.colors.atmo4}`,
-    };
-
-    return (
-      <HvTableContainer>
-        <HvTable>
-          <HvTableHead>
-            <HvTableRow>
-              {columns.map((el, index) => (
-                <HvTableHeader
-                  key={el.Header}
-                  {...(index === 0 && { ...style })}
-                >
-                  {el.Header}
-                </HvTableHeader>
-              ))}
-            </HvTableRow>
-          </HvTableHead>
-          <HvTableBody>
-            {data.map((el, index) => (
-              <HvTableRow key={el.id}>
-                {index % 3 === 0 && (
-                  <HvTableCell
-                    rowSpan={3}
-                    style={{
-                      verticalAlign: "top",
-                      ...style,
-                    }}
-                  >
-                    {el.name}
-                  </HvTableCell>
-                )}
-                <HvTableCell>{el.createdDate}</HvTableCell>
-                <HvTableCell>{el.eventType}</HvTableCell>
-                <HvTableCell>{el.status}</HvTableCell>
-                <HvTableCell>{el.riskScore}</HvTableCell>
-                <HvTableCell>{el.severity}</HvTableCell>
-                <HvTableCell>{el.priority}</HvTableCell>
-              </HvTableRow>
-            ))}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-    );
-  },
+  render: () => <GroupedRowsStory />,
 };
 
-export const ResponsiveTable = () => {
-  const columns = useMemo(() => getColumns(), []);
-  const data = useMemo(() => makeData(20), []);
-  const muiTheme = useTheme();
-
-  return (
-    <StyledResponsiveTableContainer>
-      <StyledResponsiveTable component="div">
-        <StyledResponsiveHead $breakpoints={muiTheme.breakpoints}>
-          <StyledResponsiveTableRow $breakpoints={muiTheme.breakpoints}>
-            {columns.map((el) => (
-              <StyledResponsiveTableHeader key={el.Header}>
-                <HvOverflowTooltip data={el.Header}>
-                  {el.Header}
-                </HvOverflowTooltip>
-              </StyledResponsiveTableHeader>
-            ))}
-          </StyledResponsiveTableRow>
-        </StyledResponsiveHead>
-        <StyledResponsiveBody tabIndex={0} $breakpoints={muiTheme.breakpoints}>
-          {data.map((row) => {
-            return (
-              <StyledResponsiveTableRow
-                key={row.id}
-                hover
-                $breakpoints={muiTheme.breakpoints}
-              >
-                {Object.keys(row)
-                  .slice(1)
-                  .map((key, i) => (
-                    <HvTableCell
-                      key={`${row[key]}_${columns[i].Header}`}
-                      data-label={columns[i].Header}
-                    >
-                      {row[key]}
-                    </HvTableCell>
-                  ))}
-              </StyledResponsiveTableRow>
-            );
-          })}
-        </StyledResponsiveBody>
-      </StyledResponsiveTable>
-    </StyledResponsiveTableContainer>
-  );
-};
+export const ResponsiveTable = () => <ResponsiveTableStory />;
 
 ResponsiveTable.parameters = {
   docs: {
+    source: { code: ResponsiveTableRaw },
     description: {
       story:
         "A table with non-table elements and responsive layout (try resizing your browser).",
@@ -414,78 +230,11 @@ ResponsiveTable.parameters = {
   },
 };
 
-export const ListRow = () => {
-  const [checkedIdx, toggleChecked] = useToggleIndex(0);
-
-  const columns = useMemo<HvTableColumnConfig<AssetEvent, string>[]>(() => {
-    return [...getColumns(), { Header: "Details", accessor: "link" }];
-  }, []);
-
-  const data = useMemo(
-    () => makeData(4).map((row) => ({ ...row, link: "/details" })),
-    []
-  );
-
-  return (
-    data && (
-      <HvTableContainer style={{ padding: "2px" }}>
-        <HvTable variant="listrow">
-          <HvTableHead>
-            <HvTableRow>
-              <HvTableCell variant="listcheckbox" />
-              {columns.map((el) => (
-                <HvTableHeader key={el.Header}>{el.Header}</HvTableHeader>
-              ))}
-            </HvTableRow>
-          </HvTableHead>
-          <HvTableBody withNavigation>
-            {data.map((el, idx) => {
-              return (
-                <HvTableRow key={el.id} hover selected={checkedIdx === idx}>
-                  <HvTableCell variant="listcheckbox">
-                    <HvCheckBox
-                      aria-label="Tick to select the row"
-                      checked={checkedIdx === idx}
-                      onClick={toggleChecked(idx)}
-                    />
-                  </HvTableCell>
-                  <HvTableCell>{el.name}</HvTableCell>
-                  <HvTableCell>{el.createdDate}</HvTableCell>
-                  <HvTableCell>{el.eventType}</HvTableCell>
-                  <HvTableCell>{el.status}</HvTableCell>
-                  <HvTableCell align="center">{el.riskScore}</HvTableCell>
-                  <HvTableCell>{el.severity}</HvTableCell>
-                  <HvTableCell>{el.priority}</HvTableCell>
-                  <HvTableCell variant="listactions">
-                    <HvButton
-                      variant="secondaryGhost"
-                      onClick={() => alert("CLICK!")}
-                    >
-                      View
-                    </HvButton>
-                    <HvDropDownMenu
-                      keepOpened={false}
-                      placement="left"
-                      onClick={(e, item) => alert(item.id)}
-                      dataList={[
-                        { id: "share", label: "Share" },
-                        { id: "hide", label: "Hide" },
-                        { id: "remove", label: "Remove" },
-                      ]}
-                    />
-                  </HvTableCell>
-                </HvTableRow>
-              );
-            })}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-    )
-  );
-};
+export const ListRow = () => <ListRowStory />;
 
 ListRow.parameters = {
   docs: {
+    source: { code: ListRowRaw },
     description: {
       story: "List row variant of the table.",
     },
