+++ packages/core/src/components/Table/stories/TableHooks/AlternativeLayout.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Story } from "@storybook/addon-docs";
-import { AlternativeLayoutStory } from "../TableSamples/TableSamples";
+import { AlternativeLayoutStory } from "./TableHooks";
 
 <Meta title="Visualizations/Table/Table Hooks/Alternative Layout" />
