+++ packages/core/src/components/Table/stories/TableHooks/AlternativeLayout.tsx
@@ -1,0 +1,195 @@
+import { useMemo, useState } from "react";
+import { useAbsoluteLayout, useBlockLayout, useFlexLayout } from "react-table";
+import { Delete } from "@hitachivantara/uikit-react-icons";
+import {
+  HvTable,
+  HvTableBody,
+  HvTableCell,
+  HvTableColumnConfig,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableRow,
+  useHvBulkActions,
+  useHvData,
+  useHvPagination,
+  useHvRowSelection,
+  HvBulkActions,
+  HvButton,
+  HvDropdown,
+  HvListValue,
+  HvPagination,
+  HvSwitch,
+  HvTypography,
+} from "@hitachivantara/uikit-react-core";
+
+import { AssetEvent, makeData } from "../storiesUtils";
+
+const SampleTable = ({ columns, data, layoutHook, component }) => {
+  const {
+    getTableProps,
+    getTableBodyProps,
+    prepareRow,
+    headerGroups,
+    page,
+    rows,
+    selectedFlatRows,
+    getHvBulkActionsProps,
+    getHvPaginationProps,
+  } = useHvData(
+    {
+      columns,
+      data,
+    },
+    layoutHook,
+    useHvPagination,
+    useHvRowSelection,
+    useHvBulkActions
+  );
+
+  return (
+    <>
+      <HvBulkActions
+        {...getHvBulkActionsProps?.()}
+        numTotal={rows.length}
+        numSelected={selectedFlatRows.length}
+        showSelectAllPages
+      />
+      <HvTableContainer>
+        <HvTable {...getTableProps()} component={component}>
+          <HvTableHead>
+            {headerGroups.map((headerGroup) => (
+              <HvTableRow {...headerGroup.getHeaderGroupProps()}>
+                {headerGroup.headers.map((col) => (
+                  <HvTableHeader {...col.getHeaderProps()}>
+                    {col.render("Header")}
+                  </HvTableHeader>
+                ))}
+              </HvTableRow>
+            ))}
+          </HvTableHead>
+          <HvTableBody {...getTableBodyProps()}>
+            {page.map((row) => {
+              prepareRow(row);
+              return (
+                <HvTableRow {...row.getRowProps()}>
+                  {row.cells.map((cell) => (
+                    <HvTableCell {...cell.getCellProps()}>
+                      {cell.render("Cell")}
+                    </HvTableCell>
+                  ))}
+                </HvTableRow>
+              );
+            })}
+          </HvTableBody>
+        </HvTable>
+      </HvTableContainer>
+      {page?.length ? (
+        <HvPagination {...getHvPaginationProps?.()} />
+      ) : undefined}
+    </>
+  );
+};
+
+export const AlternativeLayout = () => {
+  const alternativeLayouts = useMemo<HvListValue[]>(
+    () => [
+      {
+        id: "0",
+        label: "useFlexLayout",
+        hook: useFlexLayout,
+        selected: true,
+      },
+      { id: "1", label: "useBlockLayout", hook: useBlockLayout },
+      { id: "2", label: "useAbsoluteLayout", hook: useAbsoluteLayout },
+    ],
+    []
+  );
+
+  const [layoutHook, setLayoutHook] = useState(() => useFlexLayout);
+  const [tableElements, setTableElements] = useState(false);
+
+  const data = useMemo(() => makeData(64), []);
+
+  const columns: HvTableColumnConfig<AssetEvent, string>[] = useMemo(
+    () => [
+      { Header: "Title", accessor: "name", minWidth: 120 },
+      { Header: "Time", accessor: "createdDate", minWidth: 100 },
+      { Header: "Status", accessor: "status", width: 120 },
+      {
+        Header: "Probability",
+        accessor: "riskScore",
+        align: "right",
+        Cell: ({ value }) => <>{value}%</>,
+      },
+      { Header: "Priority", accessor: "priority" },
+      {
+        id: "actions",
+        variant: "actions",
+        width: 32,
+        Cell: () => (
+          <HvButton aria-label="Delete" icon>
+            <Delete />
+          </HvButton>
+        ),
+      },
+    ],
+    []
+  );
+
+  const table = useMemo(
+    () => (
+      <SampleTable
+        columns={columns}
+        data={data}
+        layoutHook={layoutHook}
+        component={tableElements ? "table" : "div"}
+        // Key ensures a new context for the SampleTable's
+        // useHvTable call when React reconciles the tree
+        key={layoutHook as any}
+      />
+    ),
+    [columns, data, layoutHook, tableElements]
+  );
+
+  return (
+    <>
+      <div
+        style={{ display: "flex", alignItems: "baseline", marginBottom: 20 }}
+      >
+        <div style={{ width: 200 }}>
+          <HvDropdown
+            label="Select layout"
+            values={alternativeLayouts}
+            multiSelect={false}
+            onChange={(item) =>
+              setLayoutHook(() => (item as HvListValue)?.hook)
+            }
+          />
+        </div>
+        <div style={{ display: "flex", alignItems: "center", marginLeft: 20 }}>
+          <HvTypography
+            aria-hidden="true"
+            onClick={() => setTableElements((v) => !v)}
+            style={{ marginRight: 10 }}
+          >
+            <pre>&lt;div&gt;</pre>
+          </HvTypography>
+          <HvSwitch
+            checked={tableElements}
+            aria-label="Use table html elements"
+            onChange={(_evt, newChecked) => setTableElements(newChecked)}
+          />
+          <HvTypography
+            aria-hidden="true"
+            onClick={() => setTableElements((v) => !v)}
+            style={{ marginLeft: 10 }}
+          >
+            <pre>&lt;table&gt;</pre>
+          </HvTypography>
+        </div>
+      </div>
+      {table}
+    </>
+  );
+};
