+++ packages/core/src/components/Table/stories/TableHooks/UseHvPagination.tsx
@@ -1,0 +1,82 @@
+import { useMemo } from "react";
+import range from "lodash/range";
+import {
+  HvTable,
+  HvTableBody,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableRow,
+  HvTableCell,
+  HvPagination,
+  useHvData,
+  useHvPagination,
+} from "@hitachivantara/uikit-react-core";
+
+import { makeData, getColumns, AssetEvent } from "../storiesUtils";
+
+const EmptyRow = () => (
+  <HvTableRow>
+    <HvTableCell colSpan={100} />
+  </HvTableRow>
+);
+
+export const UseHvPagination = () => {
+  const columns = useMemo(() => getColumns(), []);
+  const data = useMemo(() => makeData(32), []);
+
+  const {
+    getTableProps,
+    getTableBodyProps,
+    prepareRow,
+    headerGroups,
+    page,
+    state: { pageSize },
+    getHvPaginationProps,
+  } = useHvData<AssetEvent, string>({ columns, data }, useHvPagination);
+
+  return (
+    <>
+      <HvTableContainer>
+        <HvTable {...getTableProps()}>
+          <HvTableHead>
+            {headerGroups.map((headerGroup) => (
+              <HvTableRow {...headerGroup.getHeaderGroupProps()}>
+                {headerGroup.headers.map((col) => (
+                  <HvTableHeader {...col.getHeaderProps()}>
+                    {col.render("Header")}
+                  </HvTableHeader>
+                ))}
+              </HvTableRow>
+            ))}
+          </HvTableHead>
+          <HvTableBody {...getTableBodyProps()}>
+            {pageSize &&
+              range(pageSize).map((i) => {
+                const row = page[i];
+
+                if (!row) return <EmptyRow key={i} />;
+
+                prepareRow(row);
+
+                return (
+                  <HvTableRow {...row.getRowProps()}>
+                    {row.cells.map((cell) => {
+                      return (
+                        <HvTableCell {...cell.getCellProps()}>
+                          {cell.render("Cell")}
+                        </HvTableCell>
+                      );
+                    })}
+                  </HvTableRow>
+                );
+              })}
+          </HvTableBody>
+        </HvTable>
+      </HvTableContainer>
+      {page?.length ? (
+        <HvPagination {...getHvPaginationProps?.()} />
+      ) : undefined}
+    </>
+  );
+};
