+++ packages/core/src/components/Table/stories/TableHooks/UseHvResizeColumn.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Story } from "@storybook/addon-docs";
-import { ColumnResizeStory } from "../TableSamples/TableSamples";
+import { ColumnResizeStory } from "./TableHooks";
 
 <Meta title="Visualizations/Table/Table Hooks/useHvResizeColumn" />
