+++ packages/core/src/components/Table/stories/TableHooks/UseHvRowSelection.stories.mdx
@@ -2,8 +2,8 @@ import { Meta, Canvas, Story } from "@storybook/addon-docs";
 import {
   UseHvSelectionStory,
   UseHvSelectionControlledStory,
+  LockedSelectionStory,
 } from "./TableHooks";
-import { LockedSelectionStory } from "../TableSamples/TableSamples";
 
 <Meta title="Visualizations/Table/Table Hooks/useHvRowSelection" />
