+++ packages/core/src/components/Table/stories/TableHooks/UseHvRowSelectionLocked.tsx
@@ -1,0 +1,116 @@
+import { useMemo } from "react";
+import {
+  HvTable,
+  HvTableBody,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableRow,
+  HvTableCell,
+  HvToggleButton,
+  HvPagination,
+  HvBulkActions,
+  useHvData,
+  useHvPagination,
+  useHvRowSelection,
+  HvTableColumnConfig,
+  useHvBulkActions,
+} from "@hitachivantara/uikit-react-core";
+import { Lock, Unlock } from "@hitachivantara/uikit-react-icons";
+
+import { makeData, getColumns, AssetEvent } from "../storiesUtils";
+
+export const LockedSelection = () => {
+  const data = useMemo(() => makeData(64), []);
+
+  const columns: HvTableColumnConfig<AssetEvent, string>[] = useMemo(
+    () => [
+      ...getColumns(),
+      {
+        id: "actions",
+        variant: "actions",
+        Cell: ({ row }) => {
+          return (
+            <HvToggleButton
+              aria-label="Lock"
+              notSelectedIcon={<Unlock />}
+              selectedIcon={<Lock />}
+              selected={row.isSelectionLocked}
+              onClick={() => row.toggleRowLockedSelection?.()}
+            />
+          );
+        },
+      },
+    ],
+    []
+  );
+
+  const {
+    getTableProps,
+    getTableBodyProps,
+    prepareRow,
+    headerGroups,
+    page,
+    rows,
+    selectedFlatRows,
+    getHvBulkActionsProps,
+    getHvPaginationProps,
+  } = useHvData<AssetEvent, string>(
+    {
+      columns,
+      data,
+      aditivePageBulkSelection: true,
+      subtractivePageBulkDeselection: false,
+      initialState: {
+        selectedRowIds: { 5: true, 7: true },
+        lockedSelectionRowIds: { 2: true, 6: true },
+      },
+    },
+    useHvPagination,
+    useHvRowSelection,
+    useHvBulkActions
+  );
+
+  return (
+    <>
+      <HvBulkActions
+        {...getHvBulkActionsProps?.()}
+        numTotal={rows.length}
+        numSelected={selectedFlatRows.length}
+        showSelectAllPages
+      />
+      <HvTableContainer>
+        <HvTable {...getTableProps()}>
+          <HvTableHead>
+            {headerGroups.map((headerGroup) => (
+              <HvTableRow {...headerGroup.getHeaderGroupProps()}>
+                {headerGroup.headers.map((col) => (
+                  <HvTableHeader {...col.getHeaderProps()}>
+                    {col.render("Header")}
+                  </HvTableHeader>
+                ))}
+              </HvTableRow>
+            ))}
+          </HvTableHead>
+          <HvTableBody {...getTableBodyProps()}>
+            {page.map((row) => {
+              prepareRow(row);
+              return (
+                <HvTableRow {...row.getRowProps()}>
+                  {row.cells.map((cell) => (
+                    <HvTableCell {...cell.getCellProps()}>
+                      {cell.render("Cell")}
+                    </HvTableCell>
+                  ))}
+                </HvTableRow>
+              );
+            })}
+          </HvTableBody>
+        </HvTable>
+      </HvTableContainer>
+      {page?.length ? (
+        <HvPagination {...getHvPaginationProps?.()} />
+      ) : undefined}
+    </>
+  );
+};
